// api/mock-server/src/generated/mock-admin/contract/types.gen.ts
export type NoteDetailRecord = DomainsNotesNoteDetail & {
    id?: string;
    deletedAt?: string;
};
// ...
export type DomainsNotesNoteDetail = ({
    kind: 'basic';
} & BasicNote) | ({
    kind: 'cloze';
} & ClozeNote);

export type BasicNote = {
    deckId: Id;
    dueAt: DateTime;
    editor: BasicNoteEditor;
    id: Id;
    kind: 'basic';
    progress: number;
    reviewedAt: DateTime;
    status: NoteStatus;
    title: string;
    updatedAt: DateTime;
};

export type BasicNoteEditor = {
    back: string;
    front: string;
};

export type NoteStatus = 'in-progress' | 'mastered';

export type ClozeNote = {
    deckId: Id;
    cards: Array<ClozeNoteCard>;
    dueAt: DateTime;
    editor: ClozeNoteEditor;
    id: Id;
    kind: 'cloze';
    progress: number;
    reviewedAt: DateTime;
    status: NoteStatus;
    title: string;
    updatedAt: DateTime;
};